fix to linux

+++ src/ft_vfprintf3.c
@@ -6,7 +6,7 @@
 /*   By: tkondo <tkondo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/05 22:41:55 by tkondo            #+#    #+#             */
-/*   Updated: 2024/12/22 02:25:42 by tkondo           ###   ########.fr       */
+/*   Updated: 2024/12/23 18:48:55 by tkondo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,13 +22,13 @@ void	init_print_percent(t_format *f, t_print *p)
 void	init_print_s(t_format *f, t_print *p)
 {
 	f->s_flag.zero = 0;
-	if (!f->u_val.ptr && !(f->s_flag.period && f->prec < 6))
+	if (!f->u_val.ptr && !((size_t)f->s_flag.period && f->prec < 6))
 		f->u_val.ptr = "(null)";
 	if (!f->u_val.ptr)
 		p->inner_len = 0;
 	else
 		p->inner_len = ft_strlen(f->u_val.ptr);
-	if (f->s_flag.period && (size_t)f->prec < p->inner_len)
+	if (f->s_flag.period && f->prec < p->inner_len)
 		p->inner_len = f->prec;
 	p->p = f->u_val.ptr;
 	set_f_width(p, f);
